pyproject.toml
[tool.poetry]
name = "qaa-analysis"
version = "0.1.0"
description = "Quant address analytics – clustering + REV"
authors = ["Your Name <you@example.com>"]
packages = [{include = "qaa_analysis", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2"
pyarrow = "^15.0"
google-cloud-bigquery = "^3.19"
google-cloud-bigquery-storage = "^2.24"
scikit-learn = "^1.4"
hdbscan = {version = "^0.8", optional = true}
python-dotenv = "^1.1.0"
db-dtypes = "^1.4.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29"
pytest = "^8.3.5"
pytest-cov = "^6.0"
black = "^25.1.0"
ruff = "^0.11.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
